Issue 40: BUG: Flipped Direction Variables
Problem

It was discovered that the offset and velocity values calculated by the
flight control application are not routed/labelled correctly.

Solution

1. Clean up the flight control main loop.
2. Ensure that offset and velocity values are correctly labelled and
   routed throughout the flight control application.

Testing

Output print tests.

Issue: https://github.com/ameall/drift-fw/issues/40
